+++ shell.nix
@@ -1,6 +1,6 @@
 with import <nixpkgs> { };
 haskell.lib.buildStackProject {
-   ghc = haskell.packages.ghc822.ghc;
+   ghc = haskell.packages.ghc843.ghc;
    name = "orgstat";
    # firefox for xdg-open
    buildInputs = [ zlib git openssh gnupg gnupg1compat xdg_utils firefox gmp ];