+++ test/Test.hs
@@ -1,6 +1,6 @@
-import           Spec       (spec)
-import           Test.Hspec (hspec)
-import           Universum
+import Spec (spec)
+import Test.Hspec (hspec)
+import Universum
 
 main :: IO ()
 main = hspec spec